# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dongjing server",
        "title": "gin-base-framework Server api",
        "termsOfService": "https://github.com/hexiaoyun128",
        "contact": {},
        "license": {},
        "version": "2.0"
    },
    "paths": {
        "/auth/banner_group": {
            "post": {
                "description": "用户用户openid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "微信"
                ],
                "summary": "用户用户openid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "微信Code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/auth/user/{id}": {
            "get": {
                "description": "用户获得,只有管理员才可调用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户获得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{\"id\":1,\"name\":\"admin\",\"alias\":\"\",\"email\":\"admin@block.vc\",\"password\":\"\",\"roles\":[],\"openid\":\"\",\"active\":true,\"is_admin\":true},\"message\":\"success\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "用户信息修改,只有管理员才可调用，不修改头像",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户信息修改",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户信息ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{\"id\":1,\"name\":\"admin\",\"alias\":\"\",\"email\":\"admin@block.vc\",\"password\":\"\",\"roles\":[],\"openid\":\"\",\"active\":true,\"is_admin\":true},\"message\":\"success\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/auth/users": {
            "get": {
                "description": "用户列表，只有管理员才可调用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":[{\"id\":1,\"name\":\"admin\",\"alias\":\"\",\"email\":\"admin@block.vc\",\"password\":\"\",\"roles\":null,\"openid\":\"\",\"active\":true,\"is_admin\":true},{\"id\":2,\"name\":\"admin\",\"alias\":\"\",\"email\":\"\",\"password\":\"\",\"roles\":null,\"openid\":\"admin\",\"active\":true,\"is_admin\":false}],\"message\":\"success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}